Added Senses and Damage
Player and Goose are now damageable with virtual methods for TakeDamage and Die.
Added AreaCheck to Goose to try and hear or see the Player, set to run 10 times a second for performance.
It will wander until it detects the player. Once detected, it will follow. If you evade by leaving it's sight, it will wander again and try to Quack.